fix(assets): normalize SVG viewBox offsets and rename filter icons to avoid naming conflict
